Refactor Event display into separate component
