+++ packages/manager/src/features/NotificationCenter/PendingActions.tsx
@@ -1,40 +1,13 @@
-import { Duration } from 'luxon';
 import * as React from 'react';
-import BarPercent from 'src/components/BarPercent/BarPercent_CMR';
-import { makeStyles, Theme } from 'src/components/core/styles';
-import Typography from 'src/components/core/Typography';
 import useEvents from 'src/hooks/useEvents';
-import useLinodes from 'src/hooks/useLinodes';
-import { useTypes } from 'src/hooks/useTypes';
 import { isInProgressEvent } from 'src/store/events/event.helpers';
 import { ExtendedEvent } from 'src/store/events/event.types';
-import {
-  eventMessageGenerator,
-  eventLabelGenerator
-} from 'src/eventMessageGenerator_CMR';
-import NotificationSection, { NotificationItem } from './NotificationSection';
-import createLinkHandlerForNotification from 'src/utilities/getEventsActionLinkStrings';
-import { Link } from 'src/components/Link';
-import { formatEventSeconds } from 'src/utilities/minute-conversion/minute-conversion';
 import { reducer } from './eventQueue';
-
-const useStyles = makeStyles((theme: Theme) => ({
-  action: {
-    display: 'flex',
-    flexFlow: 'column nowrap'
-  },
-  bar: {
-    marginTop: theme.spacing()
-  }
-}));
+import NotificationSection, { NotificationItem } from './NotificationSection';
+import RenderEvent from './RenderEvent';
 
 export const PendingActions: React.FC<{}> = _ => {
-  const classes = useStyles();
   const { events } = useEvents();
-  const { linodes } = useLinodes();
-  const { types } = useTypes();
-  const _linodes = Object.values(linodes.itemsById);
-  const _types = types.entities;
 
   const [state, dispatch] = React.useReducer(reducer, {
     inProgressEvents: events.filter(isInProgressEvent),
@@ -45,65 +18,11 @@ export const PendingActions: React.FC<{}> = _ => {
     dispatch({ type: 'update', payload: { eventsFromRedux: events } });
   }, [events]);
 
-  const formatEventForDisplay = React.useCallback(
-    (event: ExtendedEvent): NotificationItem | null => {
-      const message = eventMessageGenerator(event, _linodes, _types);
-      if (message === null) {
-        return null;
-      }
-      const completed = event.percent_complete === 100;
-      const timeRemaining = event.time_remaining
-        ? ` (~${formatTimeRemaining(event.time_remaining)})`
-        : null;
-
-      const duration = formatEventSeconds(event.duration);
-
-      const linkTarget = createLinkHandlerForNotification(
-        event.action,
-        event.entity,
-        false
-      );
-      const label = linkTarget ? (
-        <Link to={linkTarget}>{eventLabelGenerator(event)}</Link>
-      ) : (
-        event.entity?.label
-      );
-      return {
-        id: `pending-action-${event.id}`,
-        body: (
-          <div className={classes.action}>
-            <Typography>
-              {label}
-              {` `}
-              {message}
-              {timeRemaining}
-              {completed
-                ? event.status === 'failed'
-                  ? ` (failed after ${duration})`
-                  : ` (completed in ${duration})`
-                : null}
-            </Typography>
-            {!completed ? (
-              <BarPercent
-                className={classes.bar}
-                max={100}
-                value={event.percent_complete ?? 0}
-                rounded
-                narrow
-              />
-            ) : null}
-          </div>
-        )
-      };
-    },
-    [classes, _linodes, _types]
-  );
-
   const allEvents = [...state.inProgressEvents, ...state.completedEvents];
 
   const actions = allEvents
     .map(formatEventForDisplay)
-    .filter(Boolean) as NotificationItem[];
+    .filter(thisAction => Boolean(thisAction.body)) as NotificationItem[];
 
   return (
     <NotificationSection
@@ -115,26 +34,9 @@ export const PendingActions: React.FC<{}> = _ => {
   );
 };
 
-export const formatTimeRemaining = (time: string) => {
-  try {
-    const [hours, minutes, seconds] = time.split(':').map(Number);
-    if (
-      [hours, minutes, seconds].some(
-        thisNumber => typeof thisNumber === 'undefined'
-      ) ||
-      [hours, minutes, seconds].some(isNaN)
-    ) {
-      // Bad input, don't display a duration
-      return null;
-    }
-    const duration = Duration.fromObject({ hours, minutes, seconds });
-    return hours > 0
-      ? `${Math.round(duration.as('hours'))} hours remaining`
-      : `${Math.round(duration.as('minutes'))} minutes remaining`;
-  } catch {
-    // Broken/unexpected input
-    return null;
-  }
-};
+const formatEventForDisplay = (event: ExtendedEvent): NotificationItem => ({
+  id: `pending-action-${event.id}`,
+  body: <RenderEvent event={event} />
+});
 
 export default React.memo(PendingActions);
