+++ packages/manager/src/features/NotificationCenter/RenderEvent.test.tsx
@@ -1,4 +1,4 @@
-import { formatTimeRemaining } from './PendingActions';
+import { formatTimeRemaining } from './RenderEvent';
 
 describe('Pending Actions', () => {
   describe('formatTimeRemaining helper', () => {
